server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepface>=0.0.95",
    "fastapi>=0.118.0",
    "flask>=3.1.2",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
    "tf-keras>=2.20.1",
    "uvicorn[standard]>=0.37.0",
    "websockets>=15.0.1",
]
